requirements.txt
torch
torchtext
torchvision
numpy
yajl
pillow
tensorboardx